requirements.txt
Flask==2.1.1
psycopg2-binary==2.9.3
plotly==5.6.0
pandas>=1.3.0